feat: added initial prototype of a table-based flux
Added a colm transformation that implements a subset of:
/docs/new_flux_ideas_and_guide.md (2020). It adds:

* @ syntax
* aggregates
* selectors first, last, min, max, top, bottom,
  in both `{}` and `fn:` syntax

This transformation is implemented as both a standalone program and as an
embedded transformation accessible from Go. The standalone program has an
accompanying shell script that can be used to transform tableflux to existing
flux and submit it to a server.